refactor: Break up insurance
